Changes to protocol.h

+++ hardware/firmware/baseStationFirmware/protocol.h
@@ -4,13 +4,13 @@
 #include <stdint.h>
 
 typedef struct __attribute__((packed)) {
-	uint16_t size;
+	uint8_t size;
 	uint16_t sourceAddr;
 	int8_t rssi;
 } rxHeader_t;
 
 typedef struct __attribute__((packed)) {
-	uint16_t size;
+	uint8_t size;
 	uint16_t destAddr;
 } txHeader_t;
